pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.1.0",
    "ruff>=0.14.10",
    "python-dotenv>=1.2.1",
    "google-search-results>=2.4.2",
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-community>=0.4.1",
    "transformers>=4.57.3",
    "openai>=2.14.0",
    "jiter>=0.12.0",
    "rich>=14.2.0",
    "google-genai>=1.56.0",
]
